[PDP simulation integration] Data type changes for price calculation

+++ features/fintech/paylater/src/main/java/com/tokopedia/pdpsimulation/paylater/viewModel/PayLaterViewModel.kt
@@ -30,7 +30,7 @@ class PayLaterViewModel @Inject constructor(
     val payLaterSimulationResultLiveData: LiveData<Result<ArrayList<PayLaterSimulationGatewayItem>>> = _payLaterSimulationResultLiveData
     var isPayLaterProductActive = false
 
-    fun getPayLaterSimulationData(amount: Int) {
+    fun getPayLaterSimulationData(amount: Long) {
         payLaterSimulationDataUseCase.cancelJobs()
         payLaterSimulationDataUseCase.getSimulationData(
                 ::onPayLaterSimulationDataSuccess,
